fix formations and angle

+++ src/main/java/com/github/manolo8/darkbot/core/entities/Ship.java
@@ -77,7 +77,7 @@ public class Ship extends Entity implements eu.darkbot.api.game.entities.Ship {
         playerInfo.update();
         attackTarget.update();
 
-        formationId = API.readInt(address, 280, 40, 56);
+        formationId = (int) API.readDouble(address, 280, 40, 56);
         invisible = API.readBoolean(API.readLong(address + 160) + 32);
 
         shipId = API.readInt(address, 192, 76);
